[GGUF FE] Native .gguf builder + stateful KV + MoE + GenAI adapter
Bring the full working GGUF frontend onto the productized port base: a native .gguf builder
(no llama.cpp dependency) that parses the container and builds a per-architecture stateful
OpenVINO graph, plus the GenAI integration to run it end-to-end.

Frontend:
- Native load path: frontend.cpp sniffs the GGUF magic and builds via TransformerBuilder ->
  GgufGraph -> GgufBuilderDecoder (node-scoped GgufDecoder). The live-decoder (llama.cpp cgraph)
  path is retained; both feed the same op translators + passes. The frontend is now discoverable
  (un-hidden) so core.read_model("model.gguf") and OpenVINO GenAI select it by extension + magic.
- Model-scope stateful/tokenizer queries ride on the InputModel facade + the decoder's model-scope
  accessors (get_model_weights / get_model_extra_inputs / get_kv_param_res_names / is_stateful /
  is_static / get_tokenizer_config); node scope stays typed and ggml-free (get_attribute /
  get_input_view_element_offset).
- Weights are surfaced uniformly as GGML_OP_NONE leaves; translate_weight handles both the builder
  (pre-extracted "gguf.blob.*") and cgraph ("data") payloads, building a compressed decompression
  subgraph (u4/u8 + zp + f16 scale), never a full f32 constant. MXFP4 MoE experts stay packed.
- Statefulness: SetRows placeholder lowered to Concat + beam_idx Gather (LowerSetRowsStateful),
  MakeStateful -> ReadValue/Assign, f16 KV cache, add_kvcache_readvalue_init, SDPA-shaped output.
  Static (NPU) models get SqueezeMatmul.
- Op table is a superset: builder ops (cos/sin/sqr/sqrt/log/top_k/elu/gelu_quick/relu/sigmoid,
  MoE moe/add_id/argsort) merged with the port's upstream ops (im2col/l2_norm/pad/ssm_conv/
  gated_delta_net/softplus). Divergent op bodies merged as supersets (rope partial+fused,
  view/reshape extra op_cases, flash_attn soft-cap+sinks, get_rows MoE gather + stateful branches).
- MoE MUL_MAT_ID dispatches: packed MXFP4 -> on-graph dequant; constant-backed experts -> fused
  ov::op::internal::GatherMatmul (compressed, memory-bounded); non-constant -> portable Gather+MatMul.
- GenAI: AdaptToGenAI pass + tokenizer rt_info so create_from_gguf runs read_model(.gguf) + adapt
  and builds the tokenizer from metadata.

CPU plugin (required for MoE compile memory):
- SnippetsMarkSkipped skips the GatherMatmul weight decompression chain.
- is_decompression_multiply recognizes GatherMatmul consumers (alongside MatMul/Convolution/
  GroupedMatMul) so MoE experts stay compressed (OLMoE q4_0 compile peak ~53GB -> ~8.9GB).

Verified end-to-end via OpenVINO GenAI (greedy_causal_lm): qwen3-0.6B, qwen2.5-0.5B, phi3-mini
(fused QKV), llama3.2-1B, TinyLlama, OLMoE-1B-7B (MoE, 8.9GB peak), gemma4-E4B (per-op rope + SWA)
all generate coherent correct text; 64/64 frontend unit tests pass.

+++ src/frontends/gguf/tests/test_ops.cpp
@@ -254,7 +254,7 @@ TEST(GGUFOps, SoftMaxAlibi) {
     const float m1 = std::pow(2.0f, -(max_bias / 2.0f) / n_head_log2);
     std::vector<float> expected(x.size());
     for (uint32_t h = 0; h < n_head; ++h) {
-        float slope = h < n_head_log2 ? std::pow(m0, static_cast<float>(h + 1)) : std::pow(m1, static_cast<float>(2 * (h - n_head_log2) + 1));
+        float slope = h < n_head_log2 ? std::pow(m0, h + 1) : std::pow(m1, 2 * (h - n_head_log2) + 1);
         for (size_t t = 0; t < T; ++t) {
             float mx = -1e30f;
             std::vector<float> z(Kd);
